[O] Ongeki: Merge repos file
