+++ src/main/java/icu/samnyan/aqua/sega/ongeki/handler/impl/GetClientBookkeepingHandler.java
@@ -2,7 +2,7 @@ package icu.samnyan.aqua.sega.ongeki.handler.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 
-import icu.samnyan.aqua.sega.ongeki.dao.userdata.UserKopRepository;
+import icu.samnyan.aqua.sega.ongeki.UserKopRepository;
 import icu.samnyan.aqua.sega.general.BaseHandler;
 import icu.samnyan.aqua.sega.util.jackson.BasicMapper;
 import org.slf4j.Logger;
