+++ src/main/java/icu/samnyan/aqua/sega/ongeki/handler/impl/GetGameRewardHandler.java
@@ -1,7 +1,7 @@
 package icu.samnyan.aqua.sega.ongeki.handler.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import icu.samnyan.aqua.sega.ongeki.dao.gamedata.GameRewardRepository;
+import icu.samnyan.aqua.sega.ongeki.GameRewardRepository;
 import icu.samnyan.aqua.sega.general.BaseHandler;
 import icu.samnyan.aqua.sega.ongeki.model.gamedata.GameReward;
 import icu.samnyan.aqua.sega.util.jackson.BasicMapper;
