+++ src/main/java/icu/samnyan/aqua/sega/ongeki/handler/impl/GetUserCardHandler.java
@@ -1,7 +1,7 @@
 package icu.samnyan.aqua.sega.ongeki.handler.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import icu.samnyan.aqua.sega.ongeki.dao.userdata.UserCardRepository;
+import icu.samnyan.aqua.sega.ongeki.UserCardRepository;
 import icu.samnyan.aqua.sega.general.BaseHandler;
 import icu.samnyan.aqua.sega.ongeki.model.userdata.UserCard;
 import icu.samnyan.aqua.sega.util.jackson.BasicMapper;
