+++ src/main/java/icu/samnyan/aqua/sega/ongeki/handler/impl/GetUserDataHandler.java
@@ -1,7 +1,7 @@
 package icu.samnyan.aqua.sega.ongeki.handler.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import icu.samnyan.aqua.sega.ongeki.dao.userdata.UserDataRepository;
+import icu.samnyan.aqua.sega.ongeki.UserDataRepository;
 import icu.samnyan.aqua.sega.general.BaseHandler;
 import icu.samnyan.aqua.sega.ongeki.model.userdata.UserData;
 import icu.samnyan.aqua.sega.util.jackson.BasicMapper;
