+++ src/main/java/icu/samnyan/aqua/sega/ongeki/handler/impl/GetUserEventMusicHandler.java
@@ -2,7 +2,7 @@ package icu.samnyan.aqua.sega.ongeki.handler.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 
-import icu.samnyan.aqua.sega.ongeki.dao.userdata.UserEventMusicRepository;
+import icu.samnyan.aqua.sega.ongeki.UserEventMusicRepository;
 import icu.samnyan.aqua.sega.general.BaseHandler;
 import icu.samnyan.aqua.sega.ongeki.model.userdata.UserEventMusic;
 import icu.samnyan.aqua.sega.util.jackson.BasicMapper;
