+++ src/main/java/icu/samnyan/aqua/sega/ongeki/handler/impl/GetUserMemoryChapterHandler.java
@@ -1,7 +1,7 @@
 package icu.samnyan.aqua.sega.ongeki.handler.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import icu.samnyan.aqua.sega.ongeki.dao.userdata.UserMemoryChapterRepository;
+import icu.samnyan.aqua.sega.ongeki.UserMemoryChapterRepository;
 import icu.samnyan.aqua.sega.general.BaseHandler;
 import icu.samnyan.aqua.sega.ongeki.model.userdata.UserMemoryChapter;
 import icu.samnyan.aqua.sega.util.jackson.BasicMapper;
