+++ src/main/java/icu/samnyan/aqua/sega/ongeki/handler/impl/GetUserMusicItemHandler.java
@@ -1,7 +1,7 @@
 package icu.samnyan.aqua.sega.ongeki.handler.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import icu.samnyan.aqua.sega.ongeki.dao.userdata.UserMusicItemRepository;
+import icu.samnyan.aqua.sega.ongeki.UserMusicItemRepository;
 import icu.samnyan.aqua.sega.general.BaseHandler;
 import icu.samnyan.aqua.sega.ongeki.model.userdata.UserMusicItem;
 import icu.samnyan.aqua.sega.util.jackson.BasicMapper;
