+++ src/main/java/icu/samnyan/aqua/sega/ongeki/handler/impl/GetUserRivalDataHandler.java
@@ -1,8 +1,8 @@
 package icu.samnyan.aqua.sega.ongeki.handler.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import icu.samnyan.aqua.sega.ongeki.dao.userdata.UserDataRepository;
-import icu.samnyan.aqua.sega.ongeki.dao.userdata.UserRivalDataRepository;
+import icu.samnyan.aqua.sega.ongeki.UserDataRepository;
+import icu.samnyan.aqua.sega.ongeki.UserRivalDataRepository;
 import icu.samnyan.aqua.sega.general.BaseHandler;
 import icu.samnyan.aqua.sega.ongeki.model.response.data.UserRivalData;
 import icu.samnyan.aqua.sega.util.jackson.BasicMapper;
