+++ src/main/java/icu/samnyan/aqua/sega/ongeki/handler/impl/GetUserTechCountHandler.java
@@ -1,7 +1,7 @@
 package icu.samnyan.aqua.sega.ongeki.handler.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import icu.samnyan.aqua.sega.ongeki.dao.userdata.UserTechCountRepository;
+import icu.samnyan.aqua.sega.ongeki.UserTechCountRepository;
 import icu.samnyan.aqua.sega.general.BaseHandler;
 import icu.samnyan.aqua.sega.ongeki.model.userdata.UserTechCount;
 import icu.samnyan.aqua.sega.util.jackson.BasicMapper;
