+++ src/main/java/icu/samnyan/aqua/sega/ongeki/handler/impl/GetUserTechEventRankingHandler.java
@@ -2,7 +2,7 @@ package icu.samnyan.aqua.sega.ongeki.handler.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 
-import icu.samnyan.aqua.sega.ongeki.dao.userdata.UserTechEventRepository;
+import icu.samnyan.aqua.sega.ongeki.UserTechEventRepository;
 import icu.samnyan.aqua.sega.general.BaseHandler;
 import icu.samnyan.aqua.sega.ongeki.model.response.data.UserTechEventRankingItem;
 import icu.samnyan.aqua.sega.ongeki.model.userdata.UserTechEvent;
