+++ src/main/java/icu/samnyan/aqua/sega/ongeki/handler/impl/UpsertUserAllHandler.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import icu.samnyan.aqua.sega.general.model.Card;
 import icu.samnyan.aqua.sega.general.model.response.UserRecentRating;
 import icu.samnyan.aqua.sega.general.service.CardService;
-import icu.samnyan.aqua.sega.ongeki.dao.userdata.*;
+import icu.samnyan.aqua.sega.ongeki.*;
 import icu.samnyan.aqua.sega.general.BaseHandler;
 import icu.samnyan.aqua.sega.ongeki.model.request.UpsertUserAll;
 import icu.samnyan.aqua.sega.ongeki.model.response.CodeResp;
